requirements.txt
PyYAML>=6.0
requests>=2.28.0
PyMuPDF>=1.23.0
arxiv
pymed
